+++ root_numpy/tmva/__init__.py
@@ -1,5 +1,5 @@
-from .factory import factory_add_events
-from .reader import reader_evaluate
+from ._factory import factory_add_events
+from ._reader import reader_evaluate
 
 
 __all__ = [
